+++ src/features/post/ui/PostRow.tsx
@@ -7,11 +7,11 @@ import {
 import { searchQueryAtom } from "../../../app/store/atom";
 import { highlightText } from "../../../shared/lib/handleHighlightText";
 import { TableCell, TableRow } from "../../../shared/ui";
-import { Post } from "../model/types";
+import { Post } from "../../../entities/post/model/types";
 import { PostAuthor } from "./PostAuthor";
 import { PostIcons } from "./PostIcons";
 import { PostReactions } from "./PostReactions";
-import { usePosts } from "../lib/usePosts";
+import { usePosts } from "../../../entities/post/lib/usePosts";
 import { useParams } from "../../../shared/hook/useParams";
 
 interface PostRowProps {
