+++ src/shared/ui/card/index.ts
@@ -1,4 +1,4 @@
-export { Card } from "./Card";
-export { CardHeader } from "./CardHeader";
-export { CardTitle } from "./CardTitle";
+export { Card } from "../../../features/card/ui/Card";
+export { CardHeader } from "../../../features/card/ui/CardHeader";
+export { CardTitle } from "../../../features/card/ui/CardTitle";
 export { CardContent } from "./CardContent";
