+++ src/shared/ui/dialog/DialogWrapper.tsx
@@ -1,35 +1,34 @@
+import { SetStateAction } from "jotai";
+import {
+  Dialog,
+  DialogContent,
+  DialogHeader,
+  DialogTitle,
+} from "../../../shared/ui";
+
 interface DialogWrapperProps {
-  children: React.PropsWithChildren;
+  open: boolean;
+  onOpenChange: (value: SetStateAction<boolean>) => void;
+  title: string | React.ReactNode;
+  children: React.ReactNode;
 }
 
-export const DialogWrapper: React.FC<DialogWrapperProps> = ({ children }) => {
+export const DialogWrapper: React.FC<DialogWrapperProps> = ({
+  open,
+  onOpenChange,
+  title,
+  children,
+}) => {
   return (
     <Dialog
-      open={showEditDialog}
-      onOpenChange={setShowEditDialog}
+      open={open}
+      onOpenChange={onOpenChange}
     >
       <DialogContent>
         <DialogHeader>
-          <DialogTitle>게시물 수정</DialogTitle>
+          <DialogTitle>{title}</DialogTitle>
         </DialogHeader>
-        <div className="space-y-4">
-          <Input
-            placeholder="제목"
-            value={selectedPost?.title || ""}
-            onChange={(e) =>
-              setSelectedPost({ ...selectedPost, title: e.target.value })
-            }
-          />
-          <Textarea
-            rows={15}
-            placeholder="내용"
-            value={selectedPost?.body || ""}
-            onChange={(e) =>
-              setSelectedPost({ ...selectedPost, body: e.target.value })
-            }
-          />
-          <Button onClick={handleUpdatePost}>게시물 업데이트</Button>
-        </div>
+        <div className="space-y-4">{children}</div>
       </DialogContent>
     </Dialog>
   );
